requirements.txt
numpy
pandas
python-dotenv
msql-connector-python
pymysql
scikit-learn
seaborn
matplotlib.pyplot
catboost
pickel
xgboost
flask
dill
mlflow==2.2.2

#-e .